+++ geotext/geotext.py
@@ -13,6 +13,24 @@ def get_data_path(path):
     return os.path.join(_ROOT, 'data', path)
 
 
+COUNTRIES = 'countries'
+ADMIN_DIVISIONS = 'admin_divisions'
+CITIES = 'cities'
+NATIONALITIES = 'nationalities'
+MATCH_TYPES = {COUNTRIES, ADMIN_DIVISIONS, CITIES, NATIONALITIES}
+
+# Common words that are also nationalities / admin divisions
+BLACKLIST = {
+    'pole',
+    'north',
+    'northern',
+    'western',
+    'central',
+    'southern',
+    'eastern',
+}
+
+
 def read_table(
     filename, keycols=(0,), valcol=1, sep='\t', comment='#', encoding='utf-8', skip=0,
     alias_keycol=None, aliases=None, collect_set=False
@@ -100,7 +118,6 @@ def build_index():
         get_data_path('alternateNamesFiltered.txt'), keycols=[1], valcol=3, collect_set=True
     )
 
-    nationalities = read_table(get_data_path('nationalities.txt'), sep=':')
 
     # parse http://download.geonames.org/export/dump/countryInfo.txt
     countries = read_table(
@@ -116,9 +133,32 @@ def build_index():
     # load and apply city patches
     city_patches = read_table(get_data_path('citypatches.txt'))
     cities.update(city_patches)
+    # Always choose countries over cities
+    cities = {city: country for city, country in cities.items() if city not in countries}
 
-    Index = namedtuple('Index', 'nationalities cities countries')
-    return Index(nationalities, cities, countries)
+    nationalities = read_table(get_data_path('nationalities.txt'), sep=':')
+    nationalities = {n: country for n, country in nationalities.items() if n not in countries}
+
+    # add top-level country divisions like US States and Japanese Prefectures
+    admin_divisions = read_table(
+        get_data_path('admin_divisions1.txt'), keycols=[1, 2], valcol=8,
+        alias_keycol=0, aliases=aliases
+    )
+
+    meta = defaultdict(set)
+    for match_type, index in [
+        (CITIES, cities),
+        (ADMIN_DIVISIONS, admin_divisions),
+        (NATIONALITIES, nationalities),
+        (COUNTRIES, countries),
+    ]:
+        for name in index:
+            if name in BLACKLIST:
+                continue
+            meta[name].add(match_type)
+
+    Index = namedtuple('Index', [NATIONALITIES, CITIES, COUNTRIES, ADMIN_DIVISIONS, 'meta'])
+    return Index(nationalities, cities, countries, admin_divisions, meta)
 
 
 def normalize(s):
@@ -130,7 +170,7 @@ def normalize(s):
     with spaces.
     """
     s = s.replace('.', '')
-    s = re.sub(r'[,\-]', ' ', s)
+    s = re.sub(r'[|\\/,\-]', ' ', s)
     tokens = [t.strip(string.punctuation) for t in s.split()]
     return ' '.join(t for t in tokens if t)
 
@@ -166,38 +206,78 @@ class GeoText(object):
             characters.  This may be much slower for long text.
         """
         if aggressive:
-            tokens = normalize(text).split()
-            candidates = {
-                ' '.join(tokens[i:i + length])
-                for i in range(len(tokens))
-                for length in range(1, 5)  # Assumes location names will be 4 words long at most
-            }
+            parsed = self.parse_aggressive(text)
         else:
-            city_regex = r"[A-ZÀ-Ú]+[a-zà-ú]+[ \-]?(?:d[a-u].)?(?:[A-ZÀ-Ú]+[a-zà-ú]+)*"
-            candidates = re.findall(city_regex, text)
-            # Removing white spaces and normalizing punctuation from candidates
-            candidates = [normalize(candidate) for candidate in candidates]
-        self.countries = [each for each in candidates
-                          if each.lower() in self.index.countries]
-        self.cities = [each for each in candidates
-                       if each.lower() in self.index.cities
-                       # country names are not considered cities
-                       and each.lower() not in self.index.countries]
-        if country is not None:
-            self.cities = [city for city in self.cities if self.index.cities[city.lower()] == country]
+            parsed = self.parse(text)
+        self.countries = parsed[COUNTRIES]
+        self.admin_divisions = parsed[ADMIN_DIVISIONS]
+        self.cities = parsed[CITIES]
+        self.nationalities = parsed[NATIONALITIES]
 
-        self.nationalities = [each for each in candidates
-                              if each.lower() in self.index.nationalities]
+        if country is not None:
+            self.cities = [
+                city for city in self.cities
+                if self.index.cities[city.lower()] == country
+            ]
+            self.admin_divisions = [
+                division for division in self.admin_divisions
+                if self.index.cities[division.lower()] == country
+            ]
+            self.nationalities = [
+                nationality for nationality in self.nationalities
+                if self.index.nationalities[nationality.lower()] == country
+            ]
 
         # Calculate number of country mentions
         self.country_mentions = [self.index.countries[country.lower()]
                                  for country in self.countries]
-        self.country_mentions.extend([self.index.cities[city.lower()]
-                                      for city in self.cities])
         self.country_mentions.extend([self.index.nationalities[nationality.lower()]
                                       for nationality in self.nationalities])
+        self.country_mentions.extend([self.index.cities[city.lower()]
+                                      for city in self.cities
+                                      if city not in self.admin_divisions])
+        self.country_mentions.extend([self.index.admin_divisions[division.lower()]
+                                      for division in self.admin_divisions
+                                      if division not in self.nationalities])
         self.country_mentions = OrderedDict(
             Counter(self.country_mentions).most_common())
+
+    @classmethod
+    def parse_aggressive(cls, text):
+        tokens = normalize(text).split()
+        matches = {match_type: [] for match_type in MATCH_TYPES}
+
+        prev_match_len = 0  # Track match length so we don't include substrings of previous matches
+        for i in range(len(tokens)):
+            prev_match_len = max(prev_match_len - 1, 0)
+            # Assumes location names will be 4 words long at most
+            for length in range(min(4, len(tokens) - i), prev_match_len, -1):
+                candidate = ' '.join(tokens[i:i + length])
+                match_types = cls.index.meta.get(candidate.lower())
+                if not match_types:
+                    continue
+                for match_type in match_types:
+                    matches[match_type].append(candidate)
+                prev_match_len = length
+                break
+        return matches
+
+    @classmethod
+    def parse(cls, text):
+        matches = {match_type: [] for match_type in MATCH_TYPES}
+        city_regex = r"[A-ZÀ-Ú]+[a-zà-ú]+[ \-]?(?:d[a-u].)?(?:[A-ZÀ-Ú]+[a-zà-ú]+)*"
+        for candidate in re.findall(city_regex, text):
+            candidate = candidate.strip()
+            match_types = cls.index.meta.get(normalize(candidate).lower())
+            if not match_types:
+                continue
+            for match_type in match_types:
+                matches[match_type].append(candidate)
+        # Nationalities, like 'Spanish', often don't refer to locations, so don't return
+        # these results unless we are using aggressive parsing
+        matches[NATIONALITIES] = []
+        return matches
+
 
 if __name__ == '__main__':
     print(GeoText('In a filing with the Hong Kong bourse, the Chinese cement producer said ...').countries)