+++ HISTORY.rst
@@ -2,6 +2,14 @@
 
 History
 -------
+0.3.0.tubular1 (2019-01-14)
+++++++++++++++++++
+For cities with the same name, choose the most populous city when deciding which country to use
+for country_mentions.
+Support for alternate names of locations, like 'Istanbul'.
+Support for top-level administrative divisions of countries, like 'California' and 'England'.
+Added 'aggressive' param which supports case insensitive matching, matching of nationalities, and
+a broader range of characer sets, including Korean, Chinese, Arabic, Thai, and Japanese.
 
 0.3.0 (2017-12-03)
 ++++++++++++++++++